+++ webapp/ui/utils.js
@@ -1,0 +1,29 @@
+export function escapeHtml(unsafe) {
+    if (typeof unsafe !== 'string') return unsafe;
+    return unsafe.replace(/&/g, "&amp;")
+        .replace(/</g, "&lt;")
+        .replace(/>/g, "&gt;")
+        .replace(/"/g, "&quot;")
+        .replace(/'/g, "&#039;");
+}
+
+export function makeUnique(filename) {
+    const parts = filename.split(".");
+    if (parts.length === 1) return `${parts[0]}_${Date.now()}`;
+    const ext = parts.pop();
+    const name = parts.join(".");
+    return `${name}_${Date.now()}.${ext}`;
+}
+
+export function isPrivateIP(ip) {
+    if (typeof ip !== 'string') return false;
+    const parts = ip.split('.').map(Number);
+    if (parts.length !== 4 || parts.some(isNaN)) return false;
+    const [p1, p2] = parts;
+    if (p1 === 10) return true;
+    if (p1 === 172 && (p2 >= 16 && p2 <= 31)) return true;
+    if (p1 === 192 && p2 === 168) return true;
+    if (p1 === 127) return true;
+    if (p1 === 169 && p2 === 254) return true;
+    return false;
+}
